LPS-75087 Document Library - Moves test class to module
