+++ modules/apps/collaboration/document-library/document-library-test/src/testIntegration/resources/com/liferay/document/library/convert/test/dependencies/liferay.jpg
